+++ login-workflow/example/src/navigation/navigation-drawer.tsx
@@ -25,6 +25,11 @@ export const navGroupItems: NavItem[] = [
         itemID: 'ResetPasswordScreenBaseExample',
         icon: { name: 'app-registration' },
     },
+    {
+        title: 'Contact',
+        itemID: 'Contact',
+        icon: { name: 'contact-page' },
+    },
 ];
 
 export type NavDrawerProps = {