+++ login-workflow/src/screens/index.ts
@@ -8,3 +8,5 @@ export * from './AccountDetailsScreen';
 export * from './CreatePasswordScreen';
 export * from './ExistingAccountSuccessScreen';
 export * from './ResetPasswordScreen';
+export * from './ChangePasswordScreen';
+export * from './ContactScreen';